Object level permissions added

+++ apps/sellers/apis.py
@@ -1,3 +1,4 @@
+from apps.sellers.permissions import CompanyPermission
 from apps.sellers.serializers import ProductSerializer, TierSerializer
 from apps.sellers.models import Product, Tier
 from rest_framework.views import APIView
@@ -11,18 +12,19 @@ from django.db import transaction
 
 
 class ProductView(APIView):
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, CompanyPermission]
 
     def get(self, request, format=None):
-        try:
+        if 'id' in request.GET:
             id=request.query_params["id"]
             # id = self.kwargs["id"]
             if id != None and Product.objects.filter(id=id).exists():
                 product = Product.objects.get(id=id)
+                self.check_object_permissions(request, product)
                 serializer= ProductSerializer(product)
             else:
                 return Response({"message": "id not recognized"})
-        except:
+        else:
             products = Product.objects.filter(company=request.user.company)
             serializer=ProductSerializer(products, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
@@ -52,7 +54,7 @@ class ProductView(APIView):
                     for tier in tiers:
                         serializerB= TierSerializer(data=tier)
                         if serializerB.is_valid():
-                            serializerB.save(product=product)
+                            serializerB.save(company=company, product=product)
                         else:
                             transaction.set_rollback(True)
                             return Response(serializerB.errors, status=status.HTTP_400_BAD_REQUEST)
@@ -67,6 +69,7 @@ class ProductView(APIView):
         id=request.query_params["id"]
         if id != None and Product.objects.filter(id=id).exists():
             obj = Product.objects.get(id=id)
+            self.check_object_permissions(request, obj)
             serializer=ProductSerializer(obj, data=data, partial=True)
             
             # old_tiers = obj.tier_set.all().values()
@@ -81,13 +84,14 @@ class ProductView(APIView):
         id=request.query_params["id"]
         if id != None and Product.objects.filter(id=id).exists():
             obj = Product.objects.get(id=id)
+            self.check_object_permissions(request, obj)
             obj.delete()
             return Response({"message": f"Product {id} has been deleted"})
         return Response({"message": "Product id not recognized"})
         
         
 class TierView(APIView):
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, CompanyPermission]
 
     
     def get(self, request, format=None):
@@ -95,6 +99,7 @@ class TierView(APIView):
             tier_id=request.query_params["tier_id"]
             if tier_id != None and Tier.objects.filter(id=tier_id).exists():
                 tier = Tier.objects.get(id=tier_id)
+                self.check_object_permissions(request, tier)
                 serializer= TierSerializer(tier)
             else:
                 return Response({"message": "tier_id not recognized"})
@@ -102,7 +107,9 @@ class TierView(APIView):
             product_id=request.query_params["product_id"]
             # id = self.kwargs["id"]
             if product_id != None and Product.objects.filter(id=product_id).exists():
-                tiers = Product.objects.get(id=product_id).tier_set.all().values()
+                product=Product.objects.get(id=product_id)
+                self.check_object_permissions(request, product)
+                tiers = p.tier_set.all().values()
                 print(tiers)
                 serializer= TierSerializer(tiers, many = True)
         
@@ -131,6 +138,7 @@ class TierView(APIView):
         id=request.query_params["id"]
         if id != None and Tier.objects.filter(id=id).exists():
             obj = Tier.objects.get(id=id)
+            self.check_object_permissions(request, obj)
             serializer=TierSerializer(obj, data=data, partial=True)
             if serializer.is_valid():
                 serializer.save()
@@ -142,6 +150,7 @@ class TierView(APIView):
         id=request.query_params["id"]
         if id != None and Tier.objects.filter(id=id).exists():
             obj = Tier.objects.get(id=id)
+            self.check_object_permissions(request, obj)
             obj.delete()
             return Response({"message": f"Tier {id} has been deleted"})
         return Response({"message": "Tier id not recognized"})
